% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.character.R
\name{as.character.soracom_operator}
\alias{as.character.soracom_group}
\alias{as.character.soracom_operator}
\alias{as.character.soracom_subscriber}
\title{Character Vectors}
\usage{
\method{as.character}{soracom_operator}(x, ...)

\method{as.character}{soracom_subscriber}(x, ...)

\method{as.character}{soracom_group}(x, ...)
}
\arguments{
\item{x}{Response by SORACOM API.}

\item{...}{Not used.}
}
\description{
Character Vectors
}

