% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_event_handlers.R
\name{build_event_handler}
\alias{build_event_handler}
\alias{change_speed_class_action}
\alias{invoke_aws_lambda_action}
\alias{rule_config}
\alias{send_mail_action}
\alias{send_mail_to_operator_action}
\title{Build Event Handlers}
\usage{
build_event_handler(target, rule, action_list, name, status = c("inactive",
  "active"), description = NA)

rule_config(type = c("DailyTrafficRule", "DailyTotalTrafficRule",
  "MonthlyTrafficRule", "MonthlyTotalTrafficRule", "CumulativeTrafficRule"),
  limit_traffic, timing = c("NEVER", "IMMEDIATELY", "BEGINNING_OF_NEXT_MONTH",
  "BEGINNING_OF_NEXT_DAY", "AFTER_ONE_DAY"))

change_speed_class_action(speed_class = c("s1.minimum", "s1.slow",
  "s1.standard", "s1.fast"), timing = c("NEVER", "IMMEDIATELY",
  "BEGINNING_OF_NEXT_MONTH", "BEGINNING_OF_NEXT_DAY", "AFTER_ONE_DAY"))

send_mail_action(to, message, title, timing = c("NEVER", "IMMEDIATELY",
  "BEGINNING_OF_NEXT_MONTH", "BEGINNING_OF_NEXT_DAY", "AFTER_ONE_DAY"))

send_mail_to_operator_action(message, title, timing = c("NEVER",
  "IMMEDIATELY", "BEGINNING_OF_NEXT_MONTH", "BEGINNING_OF_NEXT_DAY",
  "AFTER_ONE_DAY"))

invoke_aws_lambda_action(endpoint, function_name, access_key, secret_access_key,
  parameter_1, parameter_2, parameter_3, timing = c("NEVER", "IMMEDIATELY",
  "BEGINNING_OF_NEXT_MONTH", "BEGINNING_OF_NEXT_DAY", "AFTER_ONE_DAY"))
}
\arguments{
\item{target}{Target. See details.}

\item{rule}{Rule.}

\item{action_list}{Action list.}

\item{name}{Name of the handler.}

\item{status}{Handler status.}

\item{description}{Description.}

\item{type}{Configuration type.}

\item{limit_traffic}{Limit traffic in megabytes.}

\item{timing}{Timing when the handler is fired.}

\item{speed_class}{Maximum speed class.}

\item{to}{Destination of the notification mail.}

\item{message}{Message body of the notification mail.}

\item{title}{Title of the notification mail.}

\item{endpoint}{AWS Lambda's endpoint.}

\item{function_name}{AWS Lambda's function name.}

\item{access_key}{AWS access key.}

\item{secret_access_key}{AWS secret key.}

\item{parameter_1}{The first parameter of Lambda.}

\item{parameter_2}{The second parameter of Lambda.}

\item{parameter_3}{The third parameter of Lambda.}
}
\description{
Build Event Handlers
}

