% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token}
\alias{get_token}
\title{Authenticate Operator}
\usage{
get_token(email, password, timeout)
}
\arguments{
\item{email}{Your email address.}

\item{password}{Your password.}

\item{timeout}{The timeout seconds of the request token.}
}
\value{
Your API token.
}
\description{
Retrieves your SORACOM's API key.
}

