% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdata.R
\name{get_user_data}
\alias{get_user_data}
\title{Get User Data}
\usage{
get_user_data(...)
}
\arguments{
\item{...}{Handles response.
See \code{httr::content} for details.}
}
\description{
Get user data combined to the current subscriber.
}

