% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{list_groups}
\alias{create_group}
\alias{delete_group}
\alias{get_group}
\alias{list_groups}
\title{Groups Operation}
\usage{
list_groups(token)

get_group(token, group_id)

get_group(token, group_id)

create_group(token, group_name)

delete_group(token, group_id)
}
\arguments{
\item{token}{Your API token.}

\item{group_id}{The group id.}

\item{group_name}{The group name. For display on the user console.}
}
\description{
Gets, lists or creates groups operation.
}

