% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.management.R
\name{get_new_token}
\alias{change_password}
\alias{get_new_token}
\alias{get_support_token}
\title{Manage Operator}
\usage{
get_new_token(token, timeout)

change_password(token, current, new)

get_support_token(token)
}
\arguments{
\item{token}{Your API token.}

\item{timeout}{The timeout seconds of the request token.}

\item{current}{Current password.}

\item{new}{New password.}
}
\description{
Manages operators.
}
\details{
\code{new} must satisfy some criteria.
See https://console.soracom.io/#/signup for details.
}

