% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.registration.R
\name{create_operator}
\alias{create_operator}
\alias{verify_operator}
\title{Operator Registration}
\usage{
create_operator(email, password)

verify_operator(registration_token)
}
\arguments{
\item{email}{Email address.}

\item{password}{Password. See details.}

\item{registration_token}{API token sent via a verification email.}
}
\description{
Creates a new operator.
}
\details{
\code{password} must satisfy some criteria.
See https://console.soracom.io/#/signup for details.
}

