% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_filter.R
\name{query_filter}
\alias{query_filter}
\title{Query Filter}
\usage{
query_filter(tag_name, tag_value, tag_value_match_mode = c("exact", "prefix"),
  status_filter, speed_class_filter)
}
\arguments{
\item{tag_name}{Tag name for filtering.}

\item{tag_value}{Tag value for filtering.}

\item{tag_value_match_mode}{Tag match mode.}
}
\description{
Constructs a query list for request filtering.
}

