% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soracomr.R
\docType{data}
\name{rates_table}
\alias{rates_table}
\title{Table of Rates}
\format{\itemize{
\item{speed_class} Speed class.
\item{time} Time.
\item{upload_rate} Charge for uploads (JPY/MiB).
\item{download_rate} Charge for downloads (JPY/MiB).
}}
\usage{
rates_table
}
\description{
Table of Rates
}
\keyword{datasets}

