% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soracomr.R
\docType{package}
\name{soracomr}
\alias{soracomr}
\alias{soracomr-package}
\title{Management of SORACOM SIM cards}
\description{
soracomr helps you to access management api of SORACOM.
}
\section{Options}{

The following option is available:
\itemize{
\item{soracomr.language} Two-letter language (e.g. \code{"ja"}) specified in the X-Soracom-Lang header.
}
}

