% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.activation.R
\name{activate_subscriber}
\alias{activate_subscriber}
\alias{deactivate_subscriber}
\alias{disable_termination_subscriber}
\alias{enable_termination_subscriber}
\alias{terminate_subscriber}
\title{Subscriber's Activation/Termination}
\usage{
activate_subscriber(token, imsi)

deactivate_subscriber(token, imsi)

terminate_subscriber(token, imsi)

enable_termination_subscriber(token, imsi)

disable_termination_subscriber(token, imsi)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}
}
\description{
Manages the activation status of subscribers.
}

