% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.expiration.R
\name{set_expiry_time}
\alias{set_expiry_time}
\alias{unset_expiry_time}
\title{Expiration}
\usage{
set_expiry_time(token, imsi, expiration)

unset_expiry_time(token, imsi)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{expiration}{The expiry time.}
}
\description{
Sets or Unsets the expiry time of subscribers.
}

