% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.group.R
\name{attach_group}
\alias{attach_group}
\alias{detach_group}
\title{Subscriber's Group}
\usage{
attach_group(token, imsi, group_id)

detach_group(token, imsi)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{group_id}{The group name the subscriber to belong to.}
}
\description{
Subscriber's Group
}

