% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.speed.R
\name{update_speed_class}
\alias{update_speed_class}
\title{Speed}
\usage{
update_speed_class(token, imsi, class)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{class}{The speed class.}
}
\description{
Sets or Unsets the speed class of subscribers.
}

