% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{add_subscribers_tags}
\alias{add_groups_tags}
\alias{add_subscribers_tags}
\alias{delete_subscribers_tag}
\title{Tags}
\usage{
add_subscribers_tags(token, imsi, tags)

delete_subscribers_tag(token, group_id, tag_name)

add_groups_tags(token, group_id, tags)

delete_subscribers_tag(token, group_id, tag_name)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{tags}{Tags to set. See details for valid input format.}

\item{group_id}{Group ID.}

\item{tag_name}{Tag name to delete.}
}
\description{
Tags
}
\details{
\code{tags} should be one of a character vector with names, a list with names, or a data frame with column tagName and tagValue.
}

