% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_header.R
\name{to_headers}
\alias{to_headers}
\title{HTTP Request Headers}
\usage{
to_headers(token, language = getOption("soracomr.language"))
}
\arguments{
\item{token}{API Token.}

\item{lang}{Two-letter language.}
}
\description{
Constructs additional headers from the specified token.
}

