% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_handlers.R
\name{list_event_handlers}
\alias{delete_event_handler}
\alias{get_event_handler}
\alias{list_event_handlers}
\title{Event Handlers}
\usage{
list_event_handlers(token, imsi, filter)

get_event_handler(token, handler_id)

delete_event_handler(token, handler_id)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{filter}{Filter condition. If both \code{imsi} and \code{filter} are present, \code{filter} is ignored.}

\item{handler_id}{Handler ID.}
}
\description{
Get event handlers.
}
\seealso{
query_filter_event_handler
}

