% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{list_groups}
\alias{create_group}
\alias{delete_group}
\alias{get_group}
\alias{list_groups}
\title{Groups Operation}
\usage{
list_groups(token, limit, last_evaluated_key, filter)

get_group(token, group_id)

create_group(token, group_name)

delete_group(token, group_id)
}
\arguments{
\item{token}{Your API token.}

\item{limit}{Maximum number of response records.}

\item{last_evaluated_key}{The last subscriber's IMSI obtained last time.
If given, the API would return subscribers from its next.}

\item{filter}{Filter condition.}

\item{group_id}{The group id.}

\item{group_name}{The group name. For display on the user console.}
}
\description{
Gets, lists or creates groups operation.
}
\seealso{
query_filter_group
}

