% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_reset.R
\name{issue_password_reset_token}
\alias{issue_password_reset_token}
\alias{reset_password}
\title{Password Reset}
\usage{
issue_password_reset_token(email)

reset_password(token, password)
}
\arguments{
\item{email}{Email address to send a password reset token.}

\item{token}{Token sent via the reset mail.}

\item{password}{New password. See details.}
}
\description{
Password reset operations.
}
\details{
If \code{issue_password_reset_token} succeed,
then an email is send to the requested email address.
The mail cotains a token to be used in \code{reset_password}.

\code{password} must satisfy some criteria.
See https://console.soracom.io/#/signup for details.
}

