% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscribers.R
\name{list_subscribers}
\alias{get_subscriber}
\alias{list_subscribers}
\title{Subscribers Operation}
\usage{
list_subscribers(token, group_id, limit, last_evaluated_key, filter)

get_subscriber(token, imsi)
}
\arguments{
\item{token}{Your API token.}

\item{group_id}{Group's ID.}

\item{limit}{Maximum number of response records.}

\item{last_evaluated_key}{The last subscriber's IMSI obtained last time.
If given, the API would return subscribers from its next.}

\item{filter}{Filter condition. If both \code{group_id} and \code{filter} are present, \code{filter} is ignored.}

\item{imsi}{Subscriber's IMSI.}
}
\description{
Gets/Lists ubscribers.
}
\seealso{
query_filter_subscriber
}

