% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{add_subscribers_tags}
\alias{add_groups_tags}
\alias{add_subscribers_tags}
\alias{delete_groups_tag}
\alias{delete_subscribers_tag}
\alias{set_groups_name}
\alias{set_subscribers_name}
\alias{tags}
\title{Tags}
\usage{
add_subscribers_tags(token, imsi, tags)

delete_subscribers_tag(token, imsi, tag_name)

add_groups_tags(token, group_id, tags)

delete_groups_tag(token, group_id, tag_name)

set_subscribers_name(token, imsi, name)

set_groups_name(token, group_id, name)

tags(x)
}
\arguments{
\item{token}{Your API token.}

\item{imsi}{Subscriber's IMSI.}

\item{tags}{Tags to set. See details for valid input format.}

\item{tag_name}{Tag name to delete.}

\item{group_id}{Group ID.}

\item{name}{Name for subscribers or groups.}

\item{x}{Something that may have one or more tags.}
}
\description{
Tags
}
\details{
\code{tags} should be one of a character vector with names, a list with names, or a data frame with column tagName and tagValue.
}

